+++ src/compiler/compiler.rs
@@ -18,7 +18,7 @@ Parselets introduce new variable scopes.
 Loops introduce a new loop scope.
 */
 #[derive(Debug)]
-pub(super) enum Scope {
+pub(in crate::compiler) enum Scope {
     Parselet {
         // parselet-level scope (variables and constants can be defined here)
         usage_start: usize, // Begin of usages to resolve until when scope is closed
@@ -47,9 +47,9 @@ pub struct Compiler {
     parser: Option<parser::Parser>, // Internal Tokay parser
     pub debug: u8,                  // Compiler debug mode
     //values: HashSet<ImlValue>,           // Set of static intermediate values collected during compile
-    pub(super) scopes: Vec<Scope>, // Current compilation scopes
-    pub(super) usages: Vec<ImlOp>, // Unresolved calls or loads
-    pub(super) errors: Vec<Error>, // Collected errors during compilation
+    pub(in crate::compiler) scopes: Vec<Scope>, // Current compilation scopes
+    pub(in crate::compiler) usages: Vec<ImlOp>, // Unresolved calls or loads
+    pub(in crate::compiler) errors: Vec<Error>, // Collected errors during compilation
 }
 
 impl Compiler {
@@ -152,7 +152,7 @@ impl Compiler {
     }
 
     /// Tries to resolves open usages from the current scope
-    pub(super) fn resolve(&mut self) {
+    pub(in crate::compiler) fn resolve(&mut self) {
         if let Scope::Parselet { usage_start, .. } | Scope::Block { usage_start, .. } =
             &self.scopes[0]
         {
@@ -171,7 +171,7 @@ impl Compiler {
     }
 
     /// Push a parselet scope
-    pub(super) fn parselet_push(&mut self) {
+    pub(in crate::compiler) fn parselet_push(&mut self) {
         self.scopes.insert(
             0,
             Scope::Parselet {
@@ -186,7 +186,7 @@ impl Compiler {
     }
 
     /// Push a block scope
-    pub(super) fn block_push(&mut self) {
+    pub(in crate::compiler) fn block_push(&mut self) {
         self.scopes.insert(
             0,
             Scope::Block {
@@ -197,12 +197,12 @@ impl Compiler {
     }
 
     /// Push a loop scope
-    pub(super) fn loop_push(&mut self) {
+    pub(in crate::compiler) fn loop_push(&mut self) {
         self.scopes.insert(0, Scope::Loop);
     }
 
     /// Resolves and drops a parselet scope and creates a new parselet from it.
-    pub(super) fn parselet_pop(
+    pub(in crate::compiler) fn parselet_pop(
         &mut self,
         offset: Option<Offset>,
         name: Option<String>,
@@ -282,20 +282,20 @@ impl Compiler {
     }
 
     /// Drops a block scope.
-    pub(super) fn block_pop(&mut self) {
+    pub(in crate::compiler) fn block_pop(&mut self) {
         assert!(self.scopes.len() > 0 && matches!(self.scopes[0], Scope::Block { .. }));
         self.resolve();
         self.scopes.remove(0);
     }
 
     /// Drops a loop scope.
-    pub(super) fn loop_pop(&mut self) {
+    pub(in crate::compiler) fn loop_pop(&mut self) {
         assert!(self.scopes.len() > 0 && matches!(self.scopes[0], Scope::Loop));
         self.scopes.remove(0);
     }
 
     /// Marks the nearest parselet scope as consuming
-    pub(super) fn parselet_mark_consuming(&mut self) {
+    pub(in crate::compiler) fn parselet_mark_consuming(&mut self) {
         for scope in &mut self.scopes {
             if let Scope::Parselet { is_consuming, .. } = scope {
                 *is_consuming = true;
@@ -307,7 +307,7 @@ impl Compiler {
     }
 
     /// Check if there's a loop
-    pub(super) fn loop_check(&mut self) -> bool {
+    pub(in crate::compiler) fn loop_check(&mut self) -> bool {
         for i in 0..self.scopes.len() {
             match &self.scopes[i] {
                 Scope::Parselet { .. } => return false,
@@ -322,7 +322,7 @@ impl Compiler {
     /** Retrieves the address of a local variable under a given name.
 
     Returns None when the variable does not exist. */
-    pub(super) fn get_local(&self, name: &str) -> Option<usize> {
+    pub(in crate::compiler) fn get_local(&self, name: &str) -> Option<usize> {
         // Retrieve local variables from next parselet scope owning variables, except global scope!
         for scope in &self.scopes[..self.scopes.len() - 1] {
             // Check for scope with variables
@@ -339,7 +339,7 @@ impl Compiler {
     }
 
     /** Insert new local variable under given name in current scope. */
-    pub(super) fn new_local(&mut self, name: &str) -> usize {
+    pub(in crate::compiler) fn new_local(&mut self, name: &str) -> usize {
         for scope in &mut self.scopes {
             // Check for scope with variables
             if let Scope::Parselet { variables, .. } = scope {
@@ -357,7 +357,7 @@ impl Compiler {
     }
 
     /** Retrieve address of a global variable. */
-    pub(super) fn get_global(&self, name: &str) -> Option<usize> {
+    pub(in crate::compiler) fn get_global(&self, name: &str) -> Option<usize> {
         if let Scope::Parselet { variables, .. } = self.scopes.last().unwrap() {
             if let Some(addr) = variables.get(name) {
                 return Some(*addr);
@@ -370,7 +370,7 @@ impl Compiler {
     }
 
     /** Set constant to name in current scope. */
-    pub(super) fn set_constant(&mut self, name: &str, mut value: ImlValue) {
+    pub(in crate::compiler) fn set_constant(&mut self, name: &str, mut value: ImlValue) {
         /*
             Special meaning for whitespace constants names "_" and "__".
 
@@ -433,7 +433,7 @@ impl Compiler {
 
     /** Get constant value, either from current or preceding scope,
     a builtin or special. */
-    pub(super) fn get_constant(&mut self, name: &str) -> Option<ImlValue> {
+    pub(in crate::compiler) fn get_constant(&mut self, name: &str) -> Option<ImlValue> {
         // Check for constant in available scopes
         for scope in &self.scopes {
             if let Scope::Parselet { constants, .. } | Scope::Block { constants, .. } = scope {