+++ src/compiler/iml/mod.rs
@@ -6,12 +6,12 @@ mod parselet;
 mod value;
 
 use super::Linker;
-pub(crate) use op::*;
-pub(crate) use parselet::*;
-pub(super) use value::*;
+pub(in crate::compiler) use op::*;
+pub(in crate::compiler) use parselet::*;
+pub(in crate::compiler) use value::*;
 
 #[derive(Debug, Clone, PartialEq, PartialOrd)]
-pub struct Consumable {
+pub(in crate::compiler) struct Consumable {
     pub leftrec: bool,  // Flag if consumable is left-recursive
     pub nullable: bool, // Flag if consumable is nullable
 }